+++ 42-school/Common Core/Libft/ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: Dcaetano <damonteiro12446@gmail.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/26 20:09:05 by Dcaetano          #+#    #+#             */
-/*   Updated: 2023/09/28 10:42:23 by Dcaetano         ###   ########.fr       */
+/*   Updated: 2023/09/28 10:48:05 by Dcaetano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, unsigned int n)
 {
-	size_t			i;
+	unsigned int	i;
 	unsigned char	*p;
 
 	p = (unsigned char *)s;
