+++ 42-school/Common Core/Libft/libft.h
@@ -6,17 +6,15 @@
 /*   By: Dcaetano <damonteiro12446@gmail.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/26 18:36:28 by Dcaetano          #+#    #+#             */
-/*   Updated: 2023/09/28 09:32:18 by Dcaetano         ###   ########.fr       */
+/*   Updated: 2023/09/28 10:47:55 by Dcaetano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-# include <stdio.h>
-
-void	ft_bzero(void *s, size_t n);
-void	*ft_memset(void *s, int c, size_t n);
+void	ft_bzero(void *s, unsigned int n);
+void	*ft_memset(void *s, int c, unsigned int n);
 char	*ft_strrchr(const char *s, int c);
 
 #endif
